Add support for -boot flag, which allows in-tree execution of Ur/Web
The boot flag rewrites most hardcoded paths to point to the build
directory, and also forces static compilation.  This is convenient
for developing Ur/Web, or if you cannot 'sudo make install' Ur/Web.

The following changes were made:
* Header files were moved to include/urweb instead of include;
  this lets FFI users point their C_INCLUDE_PATH at this directory
  at write <urweb/urweb.h>.  For internal Ur/Web executables,
  we simply pass -I$PATH/include/urweb as normal.
* Differentiate between LIB and SRCLIB; SRCLIB is Ur and JavaScript
  source files, while LIB is compiled products from libtool.  For
  in-tree compilation these live in different places.
* No longer reference Config for paths; instead use Settings; these
  settings can be changed dynamically by Compiler.enableBoot ()
  (TODO: add a disableBoot function.)
* config.h is now generated directly in include/urweb/config.h,
  for consistency's sake (especially since it gets installed
  along with the rest of the headers!)
* All of the autotools build products got updated.
* The linkStatic field in protocols now only contains the name of the
  build product, and not the absolute path.

Future users have to be careful not to reference the Settings files
to early, lest they get an old version (this was the source of two
bugs during development of this patch.)