Fix 1-in-2^32 failure case in my_rand
If RAND_bytes happens to set ret to MIN_INT (i.e., -2^31), then
abs(ret) is undefined because 2^31 isn't representable as a 32-bit
signed integer. At least on Linux, abs(MIN_INT) returns MIN_INT, which
is still negative and causes callers to assume it failed.

This commit changes it to generate a random unsigned int, and then to
clear the top bit by shifting right 1, to ensure it's a positive value
that fits into "int".

Finally, this also moves the uw_error call into my_rand, so callers
can just assume it succeeded if it returns normally.